fixed for gwt logging

+++ src/org/adligo/i/log/client/SimpleLog.java
@@ -51,23 +51,23 @@ public class SimpleLog implements Log{
 
 
     /** "Trace" level logging. */
-    public static final int LOG_LEVEL_TRACE  = 1;
+    public static final short LOG_LEVEL_TRACE  = 1;
     /** "Debug" level logging. */
-    public static final int LOG_LEVEL_DEBUG  = 2;
+    public static final short LOG_LEVEL_DEBUG  = 2;
     /** "Info" level logging. */
-    public static final int LOG_LEVEL_INFO   = 3;
+    public static final short LOG_LEVEL_INFO   = 3;
     /** "Warn" level logging. */
-    public static final int LOG_LEVEL_WARN   = 4;
+    public static final short LOG_LEVEL_WARN   = 4;
     /** "Error" level logging. */
-    public static final int LOG_LEVEL_ERROR  = 5;
+    public static final short LOG_LEVEL_ERROR  = 5;
     /** "Fatal" level logging. */
-    public static final int LOG_LEVEL_FATAL  = 6;
+    public static final short LOG_LEVEL_FATAL  = 6;
 
     /** Enable all logging levels */
-    public static final int LOG_LEVEL_ALL    = (LOG_LEVEL_TRACE - 1);
+    public static final short LOG_LEVEL_ALL    = (LOG_LEVEL_TRACE - 1);
 
     /** Enable no logging levels */
-    public static final int LOG_LEVEL_OFF    = (LOG_LEVEL_FATAL + 1);
+    public static final short LOG_LEVEL_OFF    = (LOG_LEVEL_FATAL + 1);
 
     // ------------------------------------------------------------ Initializer
